New Test case(Create New Customer)
