+++ src/java/com/blackNecoProject/tests/NewCustomerTest.java
@@ -1,23 +1,24 @@
 package com.blackNecoProject.tests;
 
+        import com.blackNecoProject.commons.Config;
         import com.blackNecoProject.pages.LoginPage;
-        import com.blackNecoProject.pages.NewCustomer;
+        import com.blackNecoProject.pages.NewCustomerPage;
         import org.openqa.selenium.By;
         import org.testng.Assert;
         import org.testng.annotations.Test;
 
 public class NewCustomerTest extends BaseTest {
-    private NewCustomer newcustomer;
+    private NewCustomerPage newcustomerpage;
     private LoginPage loginpage;
 
     @Test(description = "Test Case 3 - Create new customer")
     public void createNewCustomer(){
-        newcustomer = new NewCustomer(driver);
+        newcustomerpage = new NewCustomerPage(driver);
         loginpage = new LoginPage(driver);
-        loginpage.loginToGuru99("mngr125317", "123!");
-        newcustomer.enterNewCustomer();
-        Assert.assertEquals(By.name("addcust"),newcustomer.FormCustomer());
-        newcustomer.fullFormNewCustomer();
-        Assert.assertEquals(By.name("regmsg"),newcustomer.getFormRegist());
+        loginpage.loginToGuru99(Config.USERNAME,Config.PASSWORD);
+        newcustomerpage.enterNewCustomer();
+        Assert.assertEquals(By.name("addcust"),newcustomerpage.FormCustomer());
+        newcustomerpage.fullFormNewCustomer();
+        Assert.assertEquals(By.name("regmsg"),newcustomerpage.getFormRegist());
     }
 }
